Add cancel command

+++ src/EggFarmSystem/EggFarmSystem.Client/Modules/MasterData/Commands/DeleteConsumableCommand.cs
@@ -1,5 +1,7 @@
-﻿using EggFarmSystem.Client.Commands;
+﻿using System.Windows;
+using EggFarmSystem.Client.Commands;
 using EggFarmSystem.Client.Core;
+using EggFarmSystem.Resources;
 using EggFarmSystem.Services;
 using System;
 using System.Collections.Generic;
@@ -20,6 +22,10 @@ namespace EggFarmSystem.Client.Modules.MasterData.Commands
 
         protected override void OnDeleteMasterData(Guid entityId)
         {
+            if(MessageBox.Show(LanguageData.General_DeleteConfirmation, LanguageData.General_Delete,MessageBoxButton.YesNo)
+                == MessageBoxResult.No)
+                return;
+
             consumableService.Delete(entityId);
         }
     }
